rep for chatting & rep decay

+++ src/Database/Models/User.cs
@@ -3,6 +3,6 @@ namespace FFA.Database.Models
     public sealed class User : Entity
     {
         public ulong UserId { get; set; }
-        public int Reputation { get; set; }
+        public double Reputation { get; set; }
     }
 }
